Removed compilation errors and warnings from main.c and calculation.c

+++ generation.c
@@ -2,16 +2,16 @@
 #include "info.h"
 
 
-int* Generate_Matrix( int matrix_width, int matrix_height )
+double* Generate_Matrix( int matrix_width, int matrix_height )
 {
-    int* output;
+    double* output;
     int arrlength = matrix_size * matrix_slice_height;
     
-    output = calloc(sizeof(int), arrlength);
+    output = calloc(sizeof(double), arrlength);
     return output;
 }
 
-void Populate_Matrix(int* arr, int length)
+void Populate_Matrix(double* arr, int length)
 {
     int x;
     for(x = 0; x<length;x++)
