+++ generation.h
@@ -4,8 +4,8 @@
 #ifndef GENERATION_H
 #define GENERATION_H
 
-int* Generate_Matrix(int matrix_width, int matrix_height);
+double* Generate_Matrix(int matrix_width, int matrix_height);
 
-void Populate_Matrix(int* arr, int length);
+void Populate_Matrix(double* arr, int length);
 
 #endif
